+++ test/tensorrt/CMakeLists.txt
@@ -11,11 +11,12 @@ if(NOT WIN32 AND TENSORRT_FOUND)
   set_tests_properties(test_converter_model_bert PROPERTIES TIMEOUT "500")
   set_tests_properties(test_converter_model_dummy PROPERTIES TIMEOUT "500")
   set_tests_properties(test_converter_model_resnet50 PROPERTIES TIMEOUT "500")
-  set_tests_properties(test_converter_conv2d PROPERTIES TIMEOUT "100")
-  set_tests_properties(test_converter_add PROPERTIES TIMEOUT "100")
-  set_tests_properties(test_converter_batch_norm PROPERTIES TIMEOUT "100")
+  set_tests_properties(test_converter_conv PROPERTIES TIMEOUT "100")
+  set_tests_properties(test_converter_norm PROPERTIES TIMEOUT "100")
   set_tests_properties(test_converter_math PROPERTIES TIMEOUT "100")
   set_tests_properties(test_converter_ops PROPERTIES TIMEOUT "100")
   set_tests_properties(test_converter_activation PROPERTIES TIMEOUT "100")
   set_tests_properties(test_converter_others PROPERTIES TIMEOUT "100")
+  set_tests_properties(test_converter_manipulation PROPERTIES TIMEOUT "100")
+  set_tests_properties(test_converter_creation PROPERTIES TIMEOUT "100")
 endif()
