+++ test/tensorrt/test_converter_math.py
@@ -82,5 +82,20 @@ class TestSubstractTRTPattern(TensorRTBaseTest):
         self.check_trt_result()
 
 
+class TestAddTRTPattern(TensorRTBaseTest):
+    def setUp(self):
+        self.python_api = paddle.add
+        self.api_args = {
+            "x": np.random.randn(2, 3).astype(np.float32),
+            "y": np.random.randn(2, 3).astype(np.float32),
+        }
+        self.program_config = {"feed_list": ["x", "y"]}
+        self.min_shape = {"x": [1, 3], "y": [1, 3]}
+        self.max_shape = {"x": [5, 3], "y": [5, 3]}
+
+    def test_trt_result(self):
+        self.check_trt_result()
+
+
 if __name__ == '__main__':
     unittest.main()
